Moveの表示関数を分離
